ISSUES-12: - Add platform-server support

+++ src/in-viewport/in-viewport.directive.ts
@@ -1,4 +1,17 @@
-import { AfterViewInit, Directive, ElementRef, EventEmitter, Input, OnDestroy, Output } from '@angular/core';
+import {
+  AfterViewInit,
+  Directive,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  Input,
+  OnDestroy,
+  Output,
+  PLATFORM_ID
+} from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Subscription } from 'rxjs/Subscription';
+
 import { InViewportConfig } from './in-viewport-config.class';
 import { InViewportService } from './in-viewport.service';
 
@@ -9,10 +22,15 @@ export const InViewportMetadata = Symbol('InViewportMetadata');
 })
 export class InViewportDirective implements AfterViewInit, OnDestroy {
   private config: InViewportConfig;
+  private subscription: Subscription = new Subscription();
 
   @Output('inViewportAction') public action$: EventEmitter<any>;
 
-  constructor(public elementRef: ElementRef, private inViewportService: InViewportService) {
+  constructor(
+    @Inject(PLATFORM_ID) private platformId: Object,
+    private elementRef: ElementRef,
+    private inViewportService: InViewportService
+  ) {
     this.config = new InViewportConfig();
     this.action$ = new EventEmitter();
   }
@@ -33,21 +51,29 @@ export class InViewportDirective implements AfterViewInit, OnDestroy {
   }
 
   ngAfterViewInit() {
-    this.inViewportService.trigger$.subscribe((entry: IntersectionObserverEntry) => this.check(entry));
-    this.inViewportService.addTarget(this.elementRef.nativeElement, this.config.rootElement);
+    if (isPlatformBrowser) {
+      this.subscription.add(
+        this.inViewportService.trigger$.subscribe((entry: IntersectionObserverEntry) => this.check(entry))
+      );
+      this.inViewportService.addTarget(this.elementRef.nativeElement, this.config.rootElement);
+    } else {
+      this.check(undefined, true);
+    }
   }
 
   ngOnDestroy() {
-    this.inViewportService.removeTarget(this.elementRef.nativeElement, this.config.rootElement);
+    if (isPlatformBrowser) {
+      this.subscription.unsubscribe();
+      this.inViewportService.removeTarget(this.elementRef.nativeElement, this.config.rootElement);
+    }
   }
 
-  check(entry: IntersectionObserverEntry) {
-    if (entry && entry.target === this.elementRef.nativeElement) {
-      const value = this.config.partial ? entry.intersectionRatio > 0 : entry.intersectionRatio === 1;
-
+  check(entry: IntersectionObserverEntry, force?: boolean) {
+    if (force || (entry && entry.target === this.elementRef.nativeElement)) {
+      const value = force || (this.config.partial ? entry.intersectionRatio > 0 : entry.intersectionRatio === 1);
       this.action$.emit({
         [InViewportMetadata]: { entry },
-        target: entry.target,
+        target: this.elementRef.nativeElement,
         value
       });
     }
